+++ CHANGELOG.md
@@ -1,11 +1,22 @@
 # Change Log
 
-## [Unreleased]
+## [0.0.5] - 2026-01-08
 
 ### Added
 
-- JSON - YAML Converter
-- Timestamp Converter
+- release workflow
+- Landing page (TBD)
+
+### Changed
+
+- Upgrade dependencies: jotai, react-router-dom, etc.
+
+## [0.0.4] - 2026-01-08
+
+### Added
+
+- compat with VSCode theme
+- release workflow (TBD)
 
 ## [0.0.3] - 2026-01-07